Crate talc

The Talc allocator crate.

For getting started:

Check out the crate’s README

Read check out the Talc and Talck structures.

Your first step will be Talc::new(...), then claim. Calling Talc::lock() on it will yield a Talck which implements [GlobalAlloc] and [Allocator] (if the appropriate feature flags are set).

Modules

locking

Note this only contains AssumeUnlockable which is not generally recommended. Use of the spin crate’s mutex with Talck is a good default.

Structs

ClaimOnOom

An out-of-memory handler that attempts to claim the memory within the given Span upon OOM.

ErrOnOom

Doesn’t handle out-of-memory conditions, immediate allocation error occurs.

Span

Represents an interval of memory [base, acme)

Talc

The Talc Allocator!

Talck

Talc lock, contains a mutex-locked Talc.

Traits

OomHandler