talc::locking

Struct AssumeUnlockable

pub struct AssumeUnlockable;

WARNING: AssumeUnlockable may cause undefined behaviour without unsafe code!

A dummy RawMutex implementation to skip synchronization on single threaded systems.

Safety

AssumeUnlockable is highly unsafe and may cause undefined behaviour if multiple threads enter a critical section it guards, even without explicit unsafe code.

Note that uncontended spin locks are cheap. Usage is only recommended on platforms that don’t have atomics or are exclusively single threaded.

Through no fault of its own, lock_api’s API does not allow for safe encapsulation of this functionality. This is a hack for backwards compatibility.