talc

Struct Span

pub struct Span { /* private fields */ }

Represents an interval of memory [base, acme)

Use get_base_acme to retrieve base and acme directly.

Empty Spans

Note that where base >= acme, the Span is considered empty, in which case the specific values of base and acme are considered meaningless.

Empty spans contain nothing and overlap with nothing.

Empty spans are contained by any sized span.

Implementations

impl Span

pub fn is_empty(self) -> bool

Returns whether base >= acme.

pub fn is_sized(self) -> bool

Returns whether base < acme.

pub fn size(self) -> usize

Returns the size of the span, else zero if base >= span.

pub fn get_base_acme(self) -> Option<(*mut u8, *mut u8)>

If self isnโt empty, returns (base, acme)

pub const fn empty() -> Self

Create an empty span.

pub const fn new(base: *mut u8, acme: *mut u8) -> Self

Create a new span.

pub const fn from_base_size(base: *mut u8, size: usize) -> Self

Creates a Span given a base and a size.

If base + size overflows, the result is empty.

pub const fn from_slice<T>(slice: *mut [T]) -> Self

pub const fn from_const_slice<T>(slice: *const [T]) -> Self

๐Deprecated: Conversion from const references encourages UB. This will be removed in a future release.

pub const fn from_array<T, const N: usize>(array: *mut [T; N]) -> Self

pub const fn from_const_array<T, const N: usize>(array: *const [T; N]) -> Self

๐Deprecated: Conversion from const references encourages UB. This will be removed in a future release.

pub fn to_ptr_range(self) -> Option<Range<*mut u8>>

Returns None if self is empty.

pub fn to_slice(self) -> Option<*mut [u8]>

Returns None if self is empty.

pub fn contains(self, ptr: *mut u8) -> bool

Returns whether self contains addr.

Empty spans contain nothing.

pub fn contains_span(self, other: Span) -> bool

Returns whether self contains other.

Empty spans are contained by any span, even empty ones.

pub fn overlaps(self, other: Span) -> bool

Returns whether some of self overlaps with other.

Empty spans donโt overlap with anything.

pub fn word_align_inward(self) -> Self

Aligns base upward and acme downward by align_of::<usize>().

pub fn word_align_outward(self) -> Self

Aligns base downward and acme upward by align_of::<usize>().

pub fn above(self, min: *mut u8) -> Self

Raises base if base is smaller than min.

pub fn below(self, max: *mut u8) -> Self

Lowers acme if acme is greater than max.

pub fn except(self, exclude: Span) -> (Self, Self)

Returns the Spans of self below and above the exclude span, respectively. Alternatively worded, the set difference self\exclude.

If exclude is empty, self and an empty Span are returned.

pub fn fit_within(self, other: Span) -> Self

Returns a span that other contains by raising base or lowering acme.

If other is empty, returns other.

pub fn fit_over(self, other: Self) -> Self

Returns a span that contains other by extending self.

If other is empty, returns self, as all spans contain any empty span.

pub fn extend(self, low: usize, high: usize) -> Self

Lower base by low and raise acme by high.

Does nothing if self is empty.

Panics

Panics if lowering base by low or raising acme by high under/overflows.

pub fn truncate(self, low: usize, high: usize) -> Span

Raise base by low and lower acme by high.

If self is empty, self is returned.

If either operation would wrap around the address space, an empty span is returned.